requirements.txt
tensorflow==2.18.0
keras==3.12.1
pillow>=10.0.0
numpy>=1.26.0
pandas
scikit-learn
streamlit==1.49.1
